+++ CHANGELOG.md
@@ -3,9 +3,9 @@ This changelog is inspired by [keepachangelog.com](https://keepachangelog.com/en
 
 ## [0.1.2.0] - 2022-02-19
 ### Added
- - 8 new voting methods:
+ - 10 new voting methods:
     - positional: borda, nauru, eurovision, dabagh, binary_positional
-    - evaluative: approval, disapproval, cav
+    - evaluative: approval, disapproval, cav, score, range
  - styling: amazing scroll shadows
 
 ### Improved